backend/requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
aiosqlite
whisper
torch
transformers
python-docx
reportlab
